Resolved 'Index Out of Bounds Error with File Filter'
Overhauled TextFileFilter class and removed charAt-based file filtering
mechanism with Apache commons-based mechanism. Usage of the Apache
commons class has resolved the out of bounds error and ensured that
encrypted files are picked up when looking for encrypted files to
decrypt. At the same, the application was altered to save an encrypted
file using a custom .enctext extension as Apache Commons IO needs to
make a check based on a file extension (and not the old _enc.txt
ending).